+++ src/cjs/VertTuple.js
@@ -233,16 +233,6 @@ var VertTuple = /** @class */ (function () {
         var t = this.getClosestT(point);
         // Compare to pointDistance?
         var distance = Math.sqrt(VertTuple.vtutils.dist2(point, this.vertAt(t)));
-        // console.log(
-        //   "distance",
-        //   distance,
-        //   "epsilon",
-        //   epsilon,
-        //   "distance < (epsilon ?? Vertex.EPSILON)",
-        //   distance < (epsilon ?? Vertex.EPSILON),
-        //   "distance < (epsilon ?? Vertex.EPSILON) && t >= 0 && t <= 1",
-        //   distance < (epsilon ?? Vertex.EPSILON) && t >= 0 && t <= 1
-        // );
         if (typeof insideBoundsOnly !== "undefined" && insideBoundsOnly) {
             return distance < (epsilon !== null && epsilon !== void 0 ? epsilon : Vertex_1.Vertex.EPSILON) && t >= 0 && t <= 1;
         }